Support/Deployment/requirements-Mac.txt
attrs==19.3.0
boto3==1.12.33
ipython==7.9.0
jedi==0.15.1
lxml==4.4.1
more-itertools==7.2.0
numpy==1.17.4
praat-parselmouth==0.3.3
psycopg2-binary==2.8.4
pycryptodome==3.9.4
pydub==0.23.1
PyQt5==5.13.1
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.1.3
QDarkStyle==2.8
sqlacodegen==2.1.0
SQLAlchemy==1.3.12
scipy==1.3.2
